+++ packages/yoga/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.115.2](https://github.com/Gympass/yoga/compare/@gympass/yoga@7.115.1...@gympass/yoga@7.115.2) (2024-06-20)
+
+**Note:** Version bump only for package @gympass/yoga
+
+
+
+
+
 ## [7.115.1](https://github.com/Gympass/yoga/compare/@gympass/yoga@7.115.0...@gympass/yoga@7.115.1) (2024-06-14)